Add error tags only for Sentry handler
